Fixed divide function in calculator class

+++ src/main/java/Calculator.java
@@ -20,8 +20,8 @@ public class Calculator {
         return result;
     }
 
-    public int divide(int number1, int number2) {
-        int result = number1 / number2;
+    public double divide(double number1, double number2) {
+        double result = number1 / number2;
         return result;
     }
 }
